\name{dataDIF}
\alias{dataDIF}
\docType{data}
\title{
	Simulated Data Set with DIF
}
\description{
	Item responses and group membership.
}
\usage{
data(dataDIF)
}
\format{
	A dataframe with 3000 dichotomous responses to 20 items (\code{I01} to \code{I20}) 
	and group membership (\code{group}).
}
\author{
	Michela Battauz
}

\seealso{
	\code{\link{dif.test}}
}
\examples{
data(dataDIF)
}
\keyword{datasets}
