\name{equateIRT-internal}
\alias{blockdiag}
\alias{irtp1}
\alias{obj}
\alias{matx}
\alias{bisco}
\alias{VarTrasf}
\alias{BisW}
\alias{VarExt}
\alias{path2link}



\title{Internal Functions of Package equateIRT}
\description{
  Internal Functions of Package equateIRT
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
