\name{equateIRT-internal}
\alias{blockdiag}
\alias{irtp1}
\alias{obj}
\alias{matx}
\alias{bisco}
\alias{VarTrasf}
\alias{BisW}
\alias{VarExt}
\alias{path2link}
\alias{fr}
\alias{truescore}
\alias{partialtruescore}
\alias{fr_cond}
\alias{DPx}
\alias{Fp}
\alias{score_prepare}
\alias{score_compute}
\alias{dif.test.compute}
\alias{trasf.compute}

\title{Internal Functions of Package equateIRT}
\description{
  Internal Functions of Package equateIRT.
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
